requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib
python-jose
python-dotenv
python-multipart
httpx
